requirements.txt
streamlit
langchain
chromadb
sentence-transformers
transformers
spacy
pandas
python-dotenv
